requirements.txt
aiohttp
fire
matplotlib
numpy==1.26.4
peft==0.10.0
torch>=2.4.0,<2.6
transformers==4.46.2
tqdm
wandb
accelerate<2
datasets
python-dotenv
openai